Added cast to match nlopt seed signature
git-svn-id: https://svn.vital-it.ch/svn/hp2c/trunk/Codeml_Variants/Fastcodeml@1797 95c5a404-1f88-0410-a6b0-c3f062b6f34a

+++ BranchSiteModel.cpp
@@ -566,7 +566,7 @@ double BranchSiteModel::maximizeLikelihood(Forest& aForest, unsigned int aFgBran
 		opt->set_lower_bounds(mLowerBound);
 		opt->set_upper_bounds(mUpperBound);
     	opt->set_ftol_abs(1e-4);
-		nlopt::srand(mSeed);
+		nlopt::srand((unsigned long)mSeed);
 
 		// If the algorithm requires a local optimizer, then add it
 		if(opt->get_algorithm() == nlopt::G_MLSL_LDS)
